+++ examples/components/arc/playback.js
@@ -1,46 +1,26 @@
 import {Arc, IconPause, IconPlay} from 'ninja';
-import React, {Component, PropTypes} from 'react';
+import React, {PropTypes} from 'react';
 
-const colorActive = 'rgb(14, 122, 254)';
+export function ArcExamplePlayback (props) {
+  const {decimal, icon, status, toggle} = props;
+  const styles = {
+    icon: {fill: icon.color, width: 32},
+    progress: {stroke: 'rgb(14, 122, 254)', strokeWidth: 6, width: 62},
+    track: {stroke: icon.color, strokeWidth: 8, width: 64}
+  };
 
-export class ArcExamplePlayback extends Component {
-  constructor (props) {
-    super(props);
-
-    this.state = {playing: true};
-  }
-
-  handleClick () {
-    const {playing} = this.state;
-
-    this.setState({playing: !playing});
-  }
-
-  render () {
-    const {icon} = this.props;
-    const {playing} = this.state;
-    const boundClick = this.handleClick.bind(this);
-    const styleIcon = {fill: icon.color, width: 32};
-
-    return (
-      <Arc onClick={boundClick} style={{
-        stroke: icon.color,
-        strokeWidth: 8,
-        width: 64
-      }}>
-        <Arc decimal={playing ? .33 : 0} style={{
-          stroke: colorActive,
-          strokeWidth: 6,
-          width: 62
-        }}>
-          {playing ?
-            <IconPause style={styleIcon}/> :
-            <IconPlay style={styleIcon}/>
+  return (
+    <div onMouseUp={toggle} onTouchEnd={toggle}>
+      <Arc style={styles.track}>
+        <Arc decimal={decimal} style={styles.progress}>
+          {status === 'playing' ?
+            <IconPause style={styles.icon}/> :
+            <IconPlay style={styles.icon}/>
           }
         </Arc>
       </Arc>
-    );
-  }
+    </div>
+  );
 }
 
 ArcExamplePlayback.propTypes = {
@@ -48,5 +28,6 @@ ArcExamplePlayback.propTypes = {
   icon: PropTypes.shape({
     color: PropTypes.string.isRequired
   }),
-  status: PropTypes.string
+  status: PropTypes.string,
+  toggle: PropTypes.func
 };