+++ examples/reducers/icon.js
@@ -1,10 +1,11 @@
-import {UPDATE_ICON_COLOR} from '../actions';
+import {iconTypes} from '../types';
+import {assign} from 'lodash';
 
 export function icon (state = {color: 'white'}, action) {
-  if (action.type === UPDATE_ICON_COLOR) {
+  if (action.type === iconTypes.UPDATE_ICON_COLOR) {
     const {color} = action;
 
-    return Object.assign({}, state, {color});
+    return assign({}, state, {color});
   }
 
   return state;