+++ examples/reducers/index.js
@@ -1,11 +1,7 @@
 import {combineReducers} from 'redux';
 import {components} from './components';
-import {download} from './download';
+import {arc} from './arc';
 import {icon} from './icon';
 import {location} from 'ruto';
-import {playback} from './playback';
-import {rainbow} from './rainbow';
 
-export const reducers = combineReducers({
-  components, download, icon, location, playback, rainbow
-});
+export const reducers = combineReducers({arc, components, icon, location});